% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mannings_to_darcy.R
\name{mannings_to_darcy}
\alias{mannings_to_darcy}
\title{Convert Manning's n to Darcy-Weisbach f}
\usage{
mannings_to_darcy(mannings, R, restrict = TRUE)
}
\arguments{
\item{mannings}{Manning's n value}

\item{R}{hydraulic radius, R in meters (m)}

\item{restrict}{allows for function parameters to restrict certain values. Type boolean. Default TRUE.}
}
\value{
Darcy-Weisbach f
}
\description{
\code{mannings_to_darcy} converts Manning's n to Darcy-Weisbach f
}
\examples{
# Result: Darcy-Weisbach f of 0.0331
mannings_to_darcy(0.030, 10)

}
